Breaks out tls-auth config
We're now configuring the tls-auth block in the client profile twice,
depending on whether unified profiles were requested, which is a
reduction in DRYness but I'd argue more readable. Besides, the config
block requires different parameters, so it makes sense to have separate
blocks.

Mostly, however, I'm doing it this way so that existing users of the
role won't see "changed" results with the same config block being moved
elsewhere.